Fixed issue with dot product still using integer math. Caused lookup to sometimes fail near the edges of triangles.

+++ users/marcel/4dworld/objects/binaural.cpp
@@ -32,7 +32,7 @@ namespace binaural
 		return result;
 	}
 	
-	int dot(const HRIRSampleGrid::CellLocation & a, const HRIRSampleGrid::CellLocation & b)
+	float dot(const HRIRSampleGrid::CellLocation & a, const HRIRSampleGrid::CellLocation & b)
 	{
 		return
 			a.elevation * b.elevation +
